Refactored code to be more elegant, added validation

+++ src/main/java/com/polinakulyk/cashregister/service/api/ReceiptService.java
@@ -1,22 +1,27 @@
 package com.polinakulyk.cashregister.service.api;
 
-import com.polinakulyk.cashregister.controller.dto.UpdateReceiptItemDto;
-import com.polinakulyk.cashregister.db.entity.Product;
 import com.polinakulyk.cashregister.db.entity.Receipt;
-import com.polinakulyk.cashregister.db.entity.ReceiptItem;
-import com.polinakulyk.cashregister.db.entity.User;
+import java.math.BigDecimal;
 import java.util.List;
-import java.util.Optional;
 
 public interface ReceiptService {
     Iterable<Receipt> findAll();
+
     List<Receipt> findAllByTellerId(String tellerId);
+
     Receipt findExistingById(String id);
+
     Receipt createReceipt(String userId);
+
     Receipt completeReceipt(String receiptId);
+
     Receipt cancelReceipt(String receiptId);
-    Receipt addReceiptItem(String receiptId, String receiptItemProductId, Integer receiptItemAmount);
+
+    Receipt addReceiptItem(
+            String receiptId, String receiptItemProductId, BigDecimal receiptItemAmount);
+
     Receipt cancelReceiptItem(String receiptId, String receiptItemId);
+
     Receipt updateReceiptItemAmount(
-            String receiptId, String receiptItemId, Integer newAmount);
+            String receiptId, String receiptItemId, BigDecimal newAmount);
 }
